refactor snapshot command & download_test
