+++ cmd/juno/juno.go
@@ -102,36 +102,7 @@ func main() {
 		return nil
 	})
 
-	var location, network string
-
-	snapshotCommand := &cobra.Command{
-		Use:   "snapshot [flags]",
-		Short: "Download snapshot based on network",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			ctx, cancel := context.WithCancel(context.Background()) //nolint
-			defer cancel()
-			go func() {
-				quiteDownload := make(chan os.Signal, 1)
-				signal.Notify(quiteDownload, os.Interrupt, syscall.SIGTERM)
-				<-quiteDownload
-				cancel()
-			}()
-			downloadError := utils.DownloadFile(ctx, network, location, &utils.RealDownloader{})
-			if downloadError != nil {
-				return downloadError
-			} else {
-				fmt.Printf("The %s snapshot was successfully downloaded to %s\nYou can run juno with the flag `--db-path` to use the downloaded snapshot\n", network, location) //nolint:lll
-			}
-			return nil
-		},
-	}
-
-	snapshotCommand.Flags().StringVar(&location, "location", "l", "Location to where the snapshot will be saved")
-	snapshotCommand.Flags().StringVar(&network, "network", "n", "Network (mainnet/goerli/goerli2)")
-	snapshotCommand.MarkFlagRequired("location") //nolint
-	snapshotCommand.MarkFlagRequired("network") //nolint
-
-	cmd.AddCommand(snapshotCommand)
+	SnapshotInit(cmd)
 	if err := cmd.ExecuteContext(ctx); err != nil {
 		os.Exit(1)
 	}
